adds admin set query and cache along with reworking HydraEndpoint for readability sake
