+++ lib/metadata.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 require 'nokogiri'
 require_relative 'loggable'
+require_relative 'nested_configuration'
 require_relative '../mapping/mapping'
-require_relative 'metadata/nested_configuration'
 require_relative 'metadata/class_method_runner'
 require_relative 'metadata/custom_node'
 require_relative 'metadata/node'
