+++ mapping/collection.rb
@@ -5,6 +5,7 @@ module Mapping
     require 'yaml'
 
     extend Extensions::BasicValueHandler
+    CSV_FILE = '../lookup/collectionlist.csv'
 
     ##
     #
@@ -13,7 +14,7 @@ module Mapping
     # @param[Array] *args -
     # return -
     def self.lookup_collection_names(value, *_args)
-      csv_file = File.join(File.dirname(__FILE__), '../lookup/collectionlist.csv')
+      csv_file = File.join(File.dirname(__FILE__), CSV_FILE)
       # array of hash
       lines = CSV.read(csv_file, headers: true).map(&:to_hash)
       lines.select { |l| value.any? { |v| l['collection_handle'].casecmp(v).zero? } }.map { |l| l['collection_name'] }
